Sessão criada

+++ src/main/java/br/ufjf/dcc193/trbo/controllers/CategoriaController.java
@@ -2,6 +2,8 @@ package br.ufjf.dcc193.trbo.controllers;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -23,46 +25,64 @@ public class CategoriaController {
 
     // CHAMA TELA CRIAR CATEGORIA
     @RequestMapping("/categoria/criar.html")
-    public String criar() {
-        return "/categoria/criar.html";
+    public String criar(HttpServletRequest request) {
+        if (request.getSession().getAttribute("atendenteLogado") != null) {
+            return "/categoria/criar.html";
+        }
+        return "redirect:/index.html";
     }
-    
+
     // CRIA CATEGORIA
     @RequestMapping(value = "/categoria/criar.html", method = RequestMethod.POST)
-    public String criar(Categoria categoria) {
-        categoriaRepo.save(categoria);
-        return "redirect:/categoria/listar.html";
+    public String criar(Categoria categoria, HttpServletRequest request) {
+        if (request.getSession().getAttribute("atendenteLogado") != null) {
+            categoriaRepo.save(categoria);
+            return "redirect:/categoria/listar.html";
+        }
+        return "redirect:/index.html";
     }
 
     // LISTA CATEGORIAS
     @RequestMapping("/categoria/listar.html")
-    public String listar(Model model) {
-        List<Categoria> listaCategorias = categoriaRepo.findAll();
-        if (listaCategorias != null) {
-            model.addAttribute("categorias", listaCategorias);
+    public String listar(Model model, HttpServletRequest request) {
+        if (request.getSession().getAttribute("atendenteLogado") != null) {
+            List<Categoria> listaCategorias = categoriaRepo.findAll();
+            if (listaCategorias != null) {
+                model.addAttribute("categorias", listaCategorias);
+            }
+            return "categoria/listar.html";
         }
-        return "categoria/listar.html";
+        return "redirect:/index.html";
     }
 
-    //CHAMA A TELA EDITAR CATEGORIA
+    // CHAMA A TELA EDITAR CATEGORIA
     @RequestMapping(value = "/categoria/editar.html/{id}")
-    public String editar(@PathVariable("id") Long id, Model model) {
-        Categoria categoria = categoriaRepo.findById(id).get();
-        model.addAttribute("categoria", categoria);
-        return "/categoria/editar.html";
+    public String editar(@PathVariable("id") Long id, Model model, HttpServletRequest request) {
+        if (request.getSession().getAttribute("atendenteLogado") != null) {
+            Categoria categoria = categoriaRepo.findById(id).get();
+            model.addAttribute("categoria", categoria);
+            return "/categoria/editar.html";
+        }
+        return "redirect:/index.html";
     }
 
-    //EDITA CATEGORIA
+    // EDITA CATEGORIA
     @RequestMapping(value = "/categoria/editar.html/{id}", method = RequestMethod.POST)
-    public String editar(Categoria categoria) {
-        categoriaRepo.save(categoria);
-        return "redirect:/categoria/listar.html";
+    public String editar(Categoria categoria, HttpServletRequest request) {
+        if (request.getSession().getAttribute("atendenteLogado") != null) {
+            categoriaRepo.save(categoria);
+            return "redirect:/categoria/listar.html";
+        }
+        return "redirect:/index.html";
     }
 
-    //DELETA CATEGORIA
+    // DELETA CATEGORIA
     @RequestMapping(value = "/categoria/deletar.html/{id}")
-    public String deletar(@PathVariable("id") Long id) {
-        categoriaRepo.deleteById(id);
-        return "redirect:/categoria/listar.html";
+    public String deletar(@PathVariable("id") Long id, HttpServletRequest request) {
+        if (request.getSession().getAttribute("atendenteLogado") != null) {
+            categoriaRepo.deleteById(id);
+            return "redirect:/categoria/listar.html";
+        }
+        return "redirect:/index.html";
     }
 }
